feat: make link icon show up on all backgrounds
